Refactored browse patch. Improves performance / reduces db overhead by only have one table for all 'item' browses (with columns for different sorting options). All sorting options are available on *all* 'item' browses - at the top level, changing a ie. from title to issue date, will make all corresponding UI hints follow. Also, maintains a separate table of all withdrawn items. This is currently not exposed in the UI, but can be easily (it's a current feature request).
git-svn-id: http://scm.dspace.org/svn/repo/trunk@2126 9c30dcfa-912a-0410-8fc2-9e0234be79fd

+++ dspace-api/src/main/java/org/dspace/browse/BrowseCreateDAOPostgres.java
@@ -40,6 +40,7 @@
 package org.dspace.browse;
 
 import java.sql.SQLException;
+import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
@@ -146,27 +147,33 @@ public class BrowseCreateDAOPostgres implements BrowseCreateDAO
     /* (non-Javadoc)
      * @see org.dspace.browse.BrowseCreateDAO#createDatabaseIndices(java.lang.String, boolean)
      */
-    public String[] createDatabaseIndices(String table, boolean execute)
+    public String[] createDatabaseIndices(String table, List<Integer> sortCols, boolean value, boolean execute)
         throws BrowseException
     {
         try
         {
-            String sortIndexName = table + "_value_index";
-            String itemIndexName = table + "_item_id_idx";
+            ArrayList<String> array = new ArrayList<String>();
             
-            String createSortIndex = "CREATE INDEX " + sortIndexName + " on " 
-                                    + table + "(sort_value);";
-            String createItemIndex = "CREATE INDEX " + itemIndexName + " ON " 
-                                    + table + "(item_id);";
+            array.add("CREATE INDEX " + table + "_item_id_idx ON " + table + "(item_id);");
+
+            if (value)
+                array.add("CREATE INDEX " + table + "_value_index ON " + table + "(sort_value);");
+
+            for (Integer i : sortCols)
+            {
+                array.add("CREATE INDEX " + table + "_s" + i + "_idx ON " + table + "(sort_" + i + ");");
+            }
             
             if (execute)
             {
-                DatabaseManager.updateQuery(context, createSortIndex);
-                DatabaseManager.updateQuery(context, createItemIndex);
+                for (String query : array)
+                {
+                    DatabaseManager.updateQuery(context, query);
+                }
             }
             
-            String[] array = { createSortIndex, createItemIndex };
-            return array;
+            String[] arr = new String[array.size()];
+            return array.toArray(arr);
         }
         catch (SQLException e)
         {
@@ -259,7 +266,44 @@ public class BrowseCreateDAOPostgres implements BrowseCreateDAO
         try
         {
             StringBuffer sb = new StringBuffer();
-            sb.append("sort_value ");
+            
+            Iterator itr = sortCols.iterator();
+            while (itr.hasNext())
+            {
+                Integer no = (Integer) itr.next();
+                sb.append(", sort_");
+                sb.append(no.toString());
+                sb.append(getSortColumnDefinition());
+            }
+            
+            String createTable = "CREATE TABLE " + table + " (" +
+                                    "id integer primary key," +
+                                    "item_id integer references item(item_id)" +
+                                    sb.toString() + 
+                                    ");";
+            if (execute)
+            {
+                DatabaseManager.updateQuery(context, createTable);
+            }
+            return createTable;
+        }
+        catch (SQLException e)
+        {
+            log.error("caught exception: ", e);
+            throw new BrowseException(e);
+        }       
+    }
+    
+    /* (non-Javadoc)
+     * @see org.dspace.browse.BrowseCreateDAO#createPrimaryTable(java.lang.String, java.util.List, boolean)
+     */
+    public String createSecondaryTable(String table, List sortCols, boolean execute)
+        throws BrowseException
+    {
+        try
+        {
+            StringBuffer sb = new StringBuffer();
+            sb.append(", sort_value ");
             sb.append(getSortColumnDefinition());
             
             Iterator itr = sortCols.iterator();
@@ -273,8 +317,8 @@ public class BrowseCreateDAOPostgres implements BrowseCreateDAO
             
             String createTable = "CREATE TABLE " + table + " (" +
                                     "id integer primary key," +
-                                    "item_id integer references item(item_id)," +
-                                    "value " + getValueColumnDefinition() + ", " + 
+                                    "item_id integer references item(item_id), " +
+                                    "value " + getValueColumnDefinition() +
                                     sb.toString() + 
                                     ");";
             if (execute)
@@ -481,6 +525,38 @@ public class BrowseCreateDAOPostgres implements BrowseCreateDAO
     /* (non-Javadoc)
      * @see org.dspace.browse.BrowseCreateDAO#insertIndex(java.lang.String, int, java.lang.String, java.lang.String, java.util.Map)
      */
+    public void insertIndex(String table, int itemID, Map sortCols)
+        throws BrowseException
+    {
+        try
+        {
+            // create us a row in the index
+            TableRow row = DatabaseManager.create(context, table);
+            
+            // set the primary information for the index
+            row.setColumn("item_id", itemID);
+            
+            // now set the columns for the other sort values
+            Iterator itra = sortCols.keySet().iterator();
+            while (itra.hasNext())
+            {
+                Integer key = (Integer) itra.next();
+                String nValue = (String) sortCols.get(key);
+                row.setColumn("sort_" + key.toString(), utils.truncateSortValue(nValue));
+            }
+            
+            DatabaseManager.update(context, row);
+        }
+        catch (SQLException e)
+        {
+            log.error("caught exception: ", e);
+            throw new BrowseException(e);
+        }
+    }
+    
+    /* (non-Javadoc)
+     * @see org.dspace.browse.BrowseCreateDAO#insertIndex(java.lang.String, int, java.lang.String, java.lang.String, java.util.Map)
+     */
     public void insertIndex(String table, int itemID, String value, String sortValue, Map sortCols)
         throws BrowseException
     {
@@ -536,22 +612,28 @@ public class BrowseCreateDAOPostgres implements BrowseCreateDAO
     /* (non-Javadoc)
      * @see org.dspace.browse.BrowseCreateDAO#pruneExcess(java.lang.String, java.lang.String)
      */
-    public void pruneExcess(String table, String map)
+    public void pruneExcess(String table, String map, boolean withdrawn)
         throws BrowseException
     {
         TableRowIterator tri = null;
         
         try
         {
-            String query = "SELECT item_id FROM " + table + " WHERE item_id NOT IN ( SELECT item_id FROM item WHERE in_archive = true AND withdrawn = false)";
+            String query = "SELECT item_id FROM " + table + " WHERE item_id NOT IN ( SELECT item_id FROM item WHERE in_archive = true AND withdrawn = " +
+                        (withdrawn ? "true" : "false") + ")";
             tri = DatabaseManager.query(context, query);
             while (tri.hasNext())
             {
                 TableRow row = tri.next();
                 String delete = "DELETE FROM " + table + " WHERE item_id = " + Integer.toString(row.getIntColumn("item_id"));
-                String deleteDistinct = "DELETE FROM " + map + " WHERE item_id = " + Integer.toString(row.getIntColumn("item_id"));
+                
                 DatabaseManager.updateQuery(context, delete);
-                DatabaseManager.updateQuery(context, deleteDistinct);
+
+                if (map != null)
+                {
+                    String deleteDistinct = "DELETE FROM " + map + " WHERE item_id = " + Integer.toString(row.getIntColumn("item_id"));
+                    DatabaseManager.updateQuery(context, deleteDistinct);
+                }
             }
         }
         catch (SQLException e)
